backend/requirements.txt
channels==3.0.4
channels-redis==4.1.0
daphne==3.0.2
Django==4.0
django-cors-headers==4.3.0
django-extra-fields==3.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
pyserial==3.5
redis==5.0.1
openpyxl==3.1.2
apsw